@@ -10,10 +10,10 @@
     "buffer": "^3.5.4",
     "jsrsasign": "5.0.1",
     "keymirror": "~0.1.0",
-    "react": "^15.1.0",
-    "react-native": "0.31",
-    "react-native-button": "*",
-    "react-native-vector-icons": "^2.1.0",
+    "react": "~15.4.0-rc.4",
+    "react-native": "0.38",
+    "react-native-button": "1.7.1",
+    "react-native-vector-icons": "4.0.0",
     "react-redux": "^4.4.5",
     "redux": "^3.4.0",
     "redux-logger": "^2.6.1",
@@ -21,9 +21,9 @@
     "redux-storage": "^4.0.0",
     "redux-storage-decorator-filter": "^1.1.4",
     "redux-storage-engine-reactnativeasyncstorage": "^1.0.0",
-    "underscore": "*",
+    "underscore": "1.8.3",
     "url": "^0.11.0",
-    "url-join": "*",
+    "url-join": "1.1.0",
     "pluralize": "3.0.0"
   },
   "devDependencies": {
